add basic ai implementation back in

+++ src/pokemon.c
@@ -2076,7 +2076,7 @@ BOOL Pokemon_TryLevelUp(struct PartyPokemon *mon) {
     }
     if (level > GetLevelCap())
         return FALSE;
-    if (exp >= GetExpByGrowthRateAndLevel((int)growthrate, level)) {
+    if ((int)exp >= GetExpByGrowthRateAndLevel((int)growthrate, level)) {
         SetMonData(mon, MON_DATA_LEVEL, &level);
         return TRUE;
     }
